% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execution_Function.R
\name{aCe_demo}
\alias{aCe_demo}
\title{Runs pre are post exercise aCe functions and prepares output overlay graph}
\usage{
aCe_demo(x, y, n)
}
\arguments{
\item{x}{pre exercise dataset created by the EndoPAT graph}

\item{y}{post exercise dataset created by the EndoPAT graph}

\item{n}{string used to name the outputs}
}
\value{
.csv files of inflection points (minima in first derivative)

.text files with description of linear regression of first portion of pulse pressure graph

PulseWave.png of Pulse pressure graph

FirstDev.png of Pulse pressure graph first derivative

InitialSlope.png of Initial linear portion of the pulse pressure graph
}
\description{
Runs pre are post exercise aCe functions and prepares output overlay graph
}
\examples{
aCe_demo("data_PreT.csv", "data_PostT.csv", "Patient_name")
}
