% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{calcLocalMax}
\alias{calcLocalMax}
\title{Execute the localMaxima Function}
\usage{
calcLocalMax(x)
}
\arguments{
\item{x}{Datatable of relative time x relative pressure}
}
\value{
MaximaRows -- Dataset of selected local maxima in ingested datatable
}
\description{
This function uses the localMaxima function to assess the data table created by read_data
}
\examples{
calcLocalMax(data2)
}
