% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{calcLocalMin}
\alias{calcLocalMin}
\title{Execute the localMaxima Function to find minima}
\usage{
calcLocalMin(x)
}
\arguments{
\item{x}{Datatable of relative time x relative pressure}
}
\value{
MaximaRows -- Dataset of local minima in ingested dataset
}
\description{
This function uses the localMaxima function to assess the ingested data table but first manipulates data (mulitply by -1) to instead find minima
}
\examples{
calcLocalMin(data2)
}
