% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callGetLLByGene.R
\name{callGetLLByGene}
\alias{callGetLLByGene}
\title{Function callGetLLByGene}
\usage{
callGetLLByGene(
  geneEss,
  useGene,
  useSamples,
  sample_effects,
  guide_efficiency,
  user_DataObj,
  user_ModelObj
)
}
\arguments{
\item{geneEss}{Double gene essentiality parameter.}

\item{useGene}{String; name of gene to evaluate under given model parameters.}

\item{useSamples}{Numeric vector; column indices of samples to evaluate.}

\item{sample_effects}{Numeric vector of sample effects.}

\item{guide_efficiency}{Numeric vector of guide efficiency parameters.}

\item{user_DataObj}{DataObj class;}

\item{user_ModelObj}{ModelObj class.}
}
\value{
returns log likelihood of given parameters, given data in parent ModelObjClass.
}
\description{
gets likelihood of gene/sample/guide parameters by calling RCpp functions
getLL, etc..  Designed to be called for a single gene.
Last updated December 2019.
}
