% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{Clientgraph}
\alias{Clientgraph}
\title{Make a ggplot overlay for client viewing}
\usage{
Clientgraph(x, y, n)
}
\arguments{
\item{x}{pre-exercise data}

\item{y}{post-exercise data}

\item{n}{string for naming}
}
\value{
ClientGraph_test.png
}
\description{
Make a ggplot overlay for client viewing
}
\examples{
Clientgraph(Pre, Post, "name", "0", "600", "0", "300")
}
