% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObjClass.R
\docType{class}
\name{DataObj}
\alias{DataObj}
\title{DataObjClass: R6 Environment Object for CRISPR Data}
\value{
DataObj(R6) of all data used in experiment.  When applicable, raw
         count data is returned.  Normalization should happen as ModelObj.
}
\description{
This function is an R6 object class to create an environment for data analysis.
Use to load all data associated with a single-target negative selection
CRISPR screen; stores all data with minimal processing, but checks for
errors and data inconsistencies.
}
\details{
All data objects in this class currently public, permitting operations without
copying over the entire dataset in function calls.
Normalization results and other preprocessing performed by ModelObjClass.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{init_counts}}{data.table of initial read counts.}

\item{\code{dep_counts}}{data.table of depleted read counts.}

\item{\code{master_counts}}{data.table of read counts from master library.}

\item{\code{neg_control_file}}{Name of file containing negative control genes.}

\item{\code{guide_covars}}{File name for sgRNA efficiency prediction features.}

\item{\code{guide2gene_map}}{Index for gene targeted by each sgRNA.}

\item{\code{init_total_reads}}{Total reads per initial sample.}

\item{\code{dep_total_reads}}{Total reads per depleted sample.}

\item{\code{cells_infected}}{Total number of cells infected per sample.}

\item{\code{sample_masterlib}}{Index for which masterlibrary went to which sample.}

\item{\code{sample_annotations}}{Sample annotations; used for partitioning test/ctrl.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataObj$new()}}
\item \href{#method-clone}{\code{DataObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataObj
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataObj$new(
  masterFiles = NA,
  countFile,
  negCtrlFile = NA,
  cellsInfected = NA,
  guideCovarFile = NA,
  sampleInfoFile = NA,
  totalReads = NA,
  hasInitSeq = T,
  sampleMasterInfoFile = NA,
  useSamples = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{masterFiles}}{Vector of file names of master libraries.  1 column sgrna.}

\item{\code{countFile}}{Text file with columns sgrna-name, gene, init_sample_1, dep_sample_1, etc.}

\item{\code{negCtrlFile}}{Text file with first column containing names of negative control genes.}

\item{\code{cellsInfected}}{Integer of total number of cells infected by masterlibrary in masterFile;
defaults to 1e3 cells per guide in master library.}

\item{\code{guideCovarFile}}{File with features to use in guide efficiency modelling.}

\item{\code{sampleInfoFile}}{File with sample initial reads info and sample types (e.g. test/ctrl)}

\item{\code{totalReads}}{Optional numeric value or vector with by-sample total read counts.}

\item{\code{hasInitSeq}}{Default True; are counts for initial infected guides
present in count file?}

\item{\code{sampleMasterInfoFile}}{File mapping samples to masterlibrary files.}

\item{\code{useSamples}}{Optional.  Sample indices in countFile to analyze.
Subsets data stored in object.
Only depleted sample indices must be given; any initial samples will be auto-matched.
 Write messages and warnings to log file.
 @param message_vector String or table to write.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
