% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuggerFunctions.R
\name{debugCpp}
\alias{debugCpp}
\title{Function debugCpp}
\usage{
debugCpp(cppOutputFile, xDataObj, xModelObj, numS = NA)
}
\arguments{
\item{cppOutputFile}{output file.}

\item{xDataObj}{DataObj used to create ModelObj & ResObj.}

\item{xModelObj}{Created by xDataObj}

\item{numS}{Number of samples; use to calculate actual number of guides from vector.}
}
\value{
List of first 10 arg indices, with $flag argument for unusual entries.
}
\description{
For debugging based on debuggging output file from optimizeModelParameters.
}
