% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuggerFunctions.R
\name{debugData}
\alias{debugData}
\title{Function debugData}
\usage{
debugData(user_DataObj, printAll = F)
}
\arguments{
\item{user_DataObj}{DataObj used to create ModelObj & ResObj.}

\item{printAll}{Boolean; whether to print all object items; else prints named subset.}
}
\value{
List of first 10 entries, with $flag argument for unusual entries.
}
\description{
Returns list of contents of DataObj.  For debugging with testthat.
These are to test for programmer, rather than user, error.
}
