% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuggerFunctions.R
\name{debugModel}
\alias{debugModel}
\title{Function debugModel}
\usage{
debugModel(user_ModelObj, user_DataObj, extraObj = c(), printAll = F)
}
\arguments{
\item{user_ModelObj}{ModelObj from DataObj.}

\item{user_DataObj}{DataObj used to create ModelObj & ResObj.}

\item{extraObj}{If you wanna throw in an extra object, go ahead.}

\item{printAll}{Boolean; whether to print all object items; else prints named subset.}
}
\value{
List of first 10 arg indices, with $flag argument for unusual entries.
}
\description{
Returns list of contents of DataObj.  For debugging with testthat.
These are to test for programmer, rather than user, error.
}
