% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debuggerFunctions.R
\name{debugRes}
\alias{debugRes}
\title{Function debugRes}
\usage{
debugRes(user_ResObj, user_ModelObj = NA, user_DataObj = NA)
}
\arguments{
\item{user_ResObj}{Object to analyze.}

\item{user_ModelObj}{ModelObj used to create ResObj}

\item{user_DataObj}{DataObj used to create ModelObj & ResObj.}
}
\value{
List of first 10 entries of objects within user_ResObj.
}
\description{
Returns list of contents of ResObj.  For debugging with testthat.
Written April 2019
Revised July 2019 to include depletion-only framework & other tweaks.
}
