% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{firstdev}
\alias{firstdev}
\title{Calculate first derivative of the pulse wave}
\usage{
firstdev(x, n)
}
\arguments{
\item{x}{data table of predicted pulse pressure values}

\item{n}{string used to name the output png}
}
\value{
dapulsewave -- data table of predicted pulse pressure values with the addition of first derivative (da) values

FirstDev_n.png
}
\description{
Calculate first derivative of the pulse wave
}
\examples{
firstdev(predpulsewave, name)
}
