% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get95Int.R
\name{get95Int}
\alias{get95Int}
\title{Function get95Int}
\usage{
get95Int(user_hessian, user_par)
}
\arguments{
\item{user_hessian}{Hessian matrix, returned from optim with hessian = T}

\item{user_par}{Value of the optimized parameter estimate; used to calculate
the confidence interval values.}
}
\value{
Returns 95% confidence interval, with respect to \code{user_par}.
}
\description{
Returns the 95% confidence interval, based on the fisher information 
derived from an optimized parameter's hessian.  
Variation is calculated as var = 1/(-Hessian) = 1/(FI), as the inverse
of the Fisher information matrix is an estimator of the asymptotic
covariance matrix.  Assuming a normal distribution (questionable),
}
