% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_nb_cpp}
\alias{get_nb_cpp}
\title{get_nb_cpp
Get negative binomial probability density for double 'count', given
size and mu parameters.}
\usage{
get_nb_cpp(mu, count, size)
}
\arguments{
\item{mu}{NumericVector}

\item{count}{double}

\item{size}{NumericVector}
}
\description{
get_nb_cpp
Get negative binomial probability density for double 'count', given
size and mu parameters.
}
