% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_norm_countVec}
\alias{get_norm_countVec}
\title{get_norm_countVec}
\usage{
get_norm_countVec(count_vector, mu, var)
}
\arguments{
\item{count_vector}{NumericVector}

\item{mu}{double}

\item{var}{double}
}
\description{
get_norm_countVec
}
