% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pois_countVec}
\alias{get_pois_countVec}
\title{get_pois_countVec
Get poisson probability density for multiple counts.}
\usage{
get_pois_countVec(count_vector, mu)
}
\arguments{
\item{count_vector}{NumericVector}

\item{mu}{double}
}
\description{
get_pois_countVec
Get poisson probability density for multiple counts.
}
