% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pois_muVec}
\alias{get_pois_muVec}
\title{get_pois_muVec}
\usage{
get_pois_muVec(count, mu_vector)
}
\arguments{
\item{count}{doubl}

\item{mu_vector}{NumericVector}
}
\description{
get_pois_muVec
}
