% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileIdx.R
\name{getFileIdx}
\alias{getFileIdx}
\title{Function getFileIdx}
\usage{
getFileIdx(baseFileName)
}
\arguments{
\item{baseFileName}{Vector of names of file to check for, for earlier runs 
with conflicting names.  Appended with '#.txt'}
}
\description{
Returns run index.
}
\examples{
getFileIdx('countData_sim_10_genes_phi_G1_v')
}
