% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGuideEfficiency.R
\name{getGuideEfficiency}
\alias{getGuideEfficiency}
\title{Function getGuideEfficiency}
\usage{
getGuideEfficiency(guide_matrix, feature_weight)
}
\arguments{
\item{guide_matrix}{Data.table of feature values for each guide.}

\item{feature_weight}{Vector of feature weights for sigmoid transform.}
}
\value{
Returns vector of by-guide percent efficiency.
}
\description{
Converts normalized feature matrix to by-guide percent efficiency.
Adds constant term not contained in guide features.
}
