% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLfcDt.R
\name{getLfcDt}
\alias{getLfcDt}
\title{getLfcDt: Function to retrieve average log fold change of count data.}
\usage{
getLfcDt(
  user_DataObj,
  isSim,
  master_freq_dt = NA,
  use_base_counts = NA,
  use_samples = NA,
  getCI = T,
  lfcFileName = "lfc",
  writeFile = F
)
}
\arguments{
\item{user_DataObj}{DataObj to analyze.}

\item{isSim}{Boolean whether data counts simulated; if so, true phi in name.}

\item{master_freq_dt}{Data.table of master library abundances if initial read counts absent.}

\item{use_base_counts}{Depleted count -matched data.table to use as the
denominator for fold change.}

\item{use_samples}{Default NA, return results for only a sample subset.}

\item{getCI}{Default true, return confidence intervals per phi.  Sim only.}

\item{lfcFileName}{Optional name of output file of lfc values.}

\item{writeFile}{Boolean default false; should output file be written.}
}
\description{
Takes geometric average across all guides and samples applying to a gene.
If negative controls provided, normalizes final averages to mean of negative
controls.  Adds a pseudocount of 0.5 to the count data after scaling all
counts to 1e7 reads per sample. Returns log2 of dep/init ratio.
}
