% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLLNoInit}
\alias{getLLNoInit}
\title{Function getLLNoInit
For experiments without sequencing of initial timepoints, just masterlib.
Removed parameters from getLL: init_counts, init_scaling.
Call get_nb and get_poiss within cpp code, returning a likelihood value to optimize.}
\usage{
getLLNoInit(
  gene_essentiality,
  guide_efficiency,
  sample_effects,
  dep_counts,
  var_model,
  master_freq,
  masterlib_key,
  cells_infected,
  dep_scaling,
  nsg_vals,
  var_params,
  step_size
)
}
\arguments{
\item{gene_essentiality}{NumericVector (for flexibility) by-guide essentiality to test.}

\item{guide_efficiency}{NumericVector by-guide efficiency.}

\item{sample_effects}{NumericVector by-sample effects on gene_essentiality.}

\item{dep_counts}{NumericMatrix}

\item{var_model}{int specifying}

\item{master_freq}{NumericMatrix; log fraction of cells infected by each guide}

\item{masterlib_key}{NumericVector of masterlibrary column to use per sample.}

\item{cells_infected}{NumericVector number of cells infected per sample in exxperiment.}

\item{dep_scaling}{NumericVector of scaling factors to normalize total
read depth and LFC ratio shifts due to guide drop-out.}

\item{nsg_vals}{Vector of percent of cells infected (unobserved, integrating over.)}

\item{var_params}{List of mean~variance model parameters (fit in ModelObj).}

\item{step_size}{Intervals of infected cells for discrete integral evaluation.}
}
\description{
Function getLLNoInit
For experiments without sequencing of initial timepoints, just masterlib.
Removed parameters from getLL: init_counts, init_scaling.
Call get_nb and get_poiss within cpp code, returning a likelihood value to optimize.
}
