% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLLNoMaster}
\alias{getLLNoMaster}
\title{Function getLLNoMaster
Version of getLL.cpp modified to return likelihood without master library
information (the infection distribution).
Call get_nb and get_poiss within cpp code, returning a likelihood value to optimize.}
\usage{
getLLNoMaster(
  gene_essentiality,
  guide_efficiency,
  sample_effects,
  init_counts,
  dep_counts,
  var_model,
  init_scaling,
  dep_scaling,
  nsg_vals,
  var_params,
  step_size
)
}
\arguments{
\item{gene_essentiality}{NumericVector (for flexibility) by-guide essentiality to test.}

\item{guide_efficiency}{NumericVector by-guide efficiency.}

\item{sample_effects}{NumericVector by-sample effects on gene_essentiality.}

\item{init_counts}{NumericMatrix}

\item{dep_counts}{NumericMatrix}

\item{var_model}{int specifying which mean-var model to use for count data.}

\item{init_scaling}{NumericVector of scaling factors to normalize total
read depth of initial sample.}

\item{dep_scaling}{NumericVector of scaling factors to normalize total
read depth and LFC ratio shifts due to guide drop-out.}

\item{nsg_vals}{Vector of percent of cells infected (unobserved, integrating over.)}

\item{var_params}{List of mean~variance model parameters (fit in ModelObj).}
}
\description{
Function getLLNoMaster
Version of getLL.cpp modified to return likelihood without master library
information (the infection distribution).
Call get_nb and get_poiss within cpp code, returning a likelihood value to optimize.
}
