% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execution_Function.R
\name{iCe_demo}
\alias{iCe_demo}
\title{Runs pre are post exercise iCe functions and prepares output overlay graph}
\usage{
iCe_demo(x, y, n, l = 0, o = 600, v = 0, z = 300)
}
\arguments{
\item{x}{pre exercise dataset created by the EndoPAT graph}

\item{y}{post exercise dataset created by the EndoPAT graph}

\item{n}{string used to name the outputs}

\item{l}{string for Pressure minima}

\item{o}{string for Pressure maxima}

\item{v}{string for Time minima}

\item{z}{string for Time maxima}
}
\value{
.csv files of inflection points (minima in first derivative)

.text files with description of linear regression of first portion of pulse pressure graph

PulseWave.png of Pulse pressure graph

FirstDev.png of Pulse pressure graph first derivative

InitialSlope.png of Initial linear portion of the pulse pressure graph
}
\description{
Runs pre are post exercise iCe functions and prepares output overlay graph
}
\examples{
iCe_demo("data_PreT.csv", "data_PostT.csv", "Patient_name")
}
