% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Execution_Function.R
\name{iCe_post}
\alias{iCe_post}
\title{Takes data extracted from EndoPAT graph to determine inputs for iCe for the pre exercise data}
\usage{
iCe_post(x, y)
}
\arguments{
\item{x}{dataset created by the EndoPAT graph}

\item{y}{string used to name the outputs}
}
\value{
.csv file of inflection points (minima in first derivative)

.text file with description of linear regression of first portion of pulse pressure graph

PulseWave.png of Pulse pressure graph

FirstDev.png of Pulse pressure graph first derivative

InitialSlope.png of Initial linear portion of the pulse pressure graph
}
\description{
Takes data extracted from EndoPAT graph to determine inputs for iCe for the pre exercise data
}
\examples{
iCe_demo("data.csv", "test")
}
