% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{iCegraph}
\alias{iCegraph}
\title{New ggplot curve to highlight key data features}
\usage{
iCegraph(x, n)
}
\arguments{
\item{x}{data table of predicted pulse pressures from the graphing function non-parametric curve}

\item{n}{string used to name the output png}
}
\value{
NewPulseWave_n.csv
}
\description{
This function uses ggplot to plot a pulse pressure graph which has in red the two inflection points and in blue the portion of the graph used to calculate the initial slope
}
\examples{
iCegraph(predpulsewave, name)

}
