% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image_Extraction.R
\name{Image_Process}
\alias{Image_Process}
\title{Image Extraction
Takes in a png image of a line graph (which is ostensibly a screenshot of EndoPat) and identifies all the pixels which are associated with a line on the line graph}
\usage{
Image_Process(x, y, z)
}
\arguments{
\item{x}{"screenshot.png"}

\item{y}{"Name_ExerciseStatus"}

\item{z}{"Treatment"}
}
\value{
Person_ExerciseStatus_TrtGroup.csv
}
\description{
Image Extraction
Takes in a png image of a line graph (which is ostensibly a screenshot of EndoPat) and identifies all the pixels which are associated with a line on the line graph
}
\examples{
Image_Process("patient.png", "AndrewGow_Pre", "T")
}
