% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{inflecpoints}
\alias{inflecpoints}
\title{Find minimum points of first derivative}
\usage{
inflecpoints(x)
}
\arguments{
\item{x}{data table of predicted pulse pressure values which includes first derivatives}
}
\value{
inflection_points -- two lowest amplitude minima of the first derivative graph
}
\description{
Finds the inflection points of the pulse pressure graph to be used in aCe
}
\examples{
inflecpoints(dapulsewave)
}
