% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{initialslope}
\alias{initialslope}
\title{Calculate slope of linear phase of pulse wave}
\usage{
initialslope(x, n)
}
\arguments{
\item{x}{data table of predicted pulse pressure values}

\item{n}{string used to name the outputs}
}
\value{
InitialSlope_n.png
}
\description{
This is used to find a linear regression of the first portion of the graph to be output for use in aCe
}
\examples{
initialslope(predpulsewave, name)
}
