% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{localMaxima}
\alias{localMaxima}
\title{Calculate the Local Maxima of a Dataset}
\usage{
localMaxima(x)
}
\arguments{
\item{x}{the dataset contaiing the dependent variable of the data set for which you want the maxima}
}
\value{
y -- vector of row numbers which describe local maxima
}
\description{
This function serves to calculate all the local maxima of a dataset or, if the data is first multiplied by -1 in the input, the local minima
}
\examples{
Output <- localMaxima(datatable$dependentvariable)
}
