% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{localvalremoveMax}
\alias{localvalremoveMax}
\title{Clean MaximaRows output to isolate valid systolic BP}
\usage{
localvalremoveMax(x)
}
\arguments{
\item{x}{Dataset of local maxima in ingested dataset}
}
\value{
FinalMaxima4 -- Dataset of cleaned local maxima (nearly completely diastolic BP) in ingested dataset
}
\description{
local maxima for a pressure wave will find systolic BP but also a point into a dichroic notch which this function will aim to remove
Note that this local value removal code serves limited function now with the new data extraction method but was left in to prevent any potential extraneous values
}
\examples{
localvalremoveMax(MaximaRows)
}
