% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{localvalremoveMin}
\alias{localvalremoveMin}
\title{Clean Minima output to isolate valid diastolic BP}
\usage{
localvalremoveMin(x)
}
\arguments{
\item{x}{Dataset of local maxima in ingested dataset}
}
\value{
FinalMaxima3 -- Dataset of cleaned local maxima in ingested dataset
}
\description{
local maxima (w/ -1 manipulation) for a pressure wave will find diastolic BP but also a point into a dichroic notch which this function will aim to remove
Note that this local value removal code serves limited function now with the new data extraction method but was left in to prevent any potential extraneous values
}
\examples{
localvalremoveMin(MinimaRows)
}
