% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelObjClass.R
\docType{class}
\name{ModelObj}
\alias{ModelObj}
\title{ModelObj Environment}
\value{
ModelObj (R6Class).
Contains:
--  mean_var_model_params = "list",
--  guide_covar = "matrix",
--  unobserved_infected_cell_values = "vector",
--  mean_var_model = "integer",
--  init_scaling = "vector",
--  dep_scaling = "vector",
--  test_sample_subtype_cols = "vector"
-- stepSize = 10
-- guidePrior = 'string'
}
\description{
ModelObj Environment

ModelObj Environment
}
\details{
This R6 environment object processes the raw data loaded in the provided
\code{'user_DataObj'} to calculate model constants according to user
specifications.  Computes sample scaling parameters, the vector of infected
 cell values to use in the likelihood summation,
 and partitions test and control samples,
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mean_var_model_params}}{Coefficients to use in mean-var model.}

\item{\code{guide_covar}}{Matrix of by-guide features.}

\item{\code{master_freq_dt}}{Data.table of sgRNA frequencies in master libraries.}

\item{\code{unobserved_infected_cell_values}}{Discrete values of infected cells for Riemann sum.}

\item{\code{mean_var_model}}{Integer specifying which form of model to use.}

\item{\code{init_scaling}}{By-sample vector of sample scaling factors.}

\item{\code{dep_scaling}}{By-sample vector of depleted sample scaling.}

\item{\code{test_sample_subtype_cols}}{Vector of column indices for 'test' subtype.}

\item{\code{stepSize}}{Integer of infected cell intervals to evaluate.}

\item{\code{guidePrior}}{String indicating form of infection prior for README.}

\item{\code{use_neg_ctrl}}{Boolean indicating whether negative controls from DataObj should be used.}

\item{\code{neg_ctrls}}{String vector of gene names for negative controls.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelObj$new()}}
\item \href{#method-clone}{\code{ModelObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create ModelObj
 Write messages and warnings to log file.
 @param message_vector String or table to write.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelObj$new(
  user_DataObj = NA,
  use_master_library = T,
  fit_guide_parameter = F,
  mean_var_model = 1,
  use_neg_ctrl = T,
  test_samples = NA,
  stepSize = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_DataObj}}{Containing all count data; a DataObj.}

\item{\code{use_master_library}}{Boolean, use masterlibrary to set guide abundance prior?}

\item{\code{fit_guide_parameter}}{Boolean, should by-guide efficiencies be calculated?}

\item{\code{mean_var_model}}{Optional integer for mean~var fitting model; default poisson.}

\item{\code{use_neg_ctrl}}{Boolean, use negative control genes for scaling sample
parameters and final essentiality value.}

\item{\code{test_samples}}{String indicating which sample annotation to use when
grouping test and control groups.}

\item{\code{stepSize}}{Number of cells to use for each step in the summation, default 10.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
