% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{nonparam}
\alias{nonparam}
\title{Calculate Pulse wave and Fit non-parametric curve}
\usage{
nonparam(x, y, n)
}
\arguments{
\item{x}{data table of systolic maxima}

\item{y}{data table of diastolic minima}

\item{n}{string used to name the output png}
}
\value{
predpulsewave -- data table of predicted pulse pressure values
}
\description{
This function calculates a non-parametric curve (spline) for the systolic and then diastolic BP data tables, predicts a set of values for each, and uses the predicted values to calculate a pulse pressure for which it then outputs a nonparametric curve
}
\examples{
nonparam(FinalMaxima3, FinalMinima3, name)
}
