% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optGuide.R
\name{optGuide}
\alias{optGuide}
\title{Function optGuide}
\usage{
optGuide(
  guide_features,
  gene_effects,
  sample_effects,
  use_genes,
  write_log,
  user_DataObj,
  user_ModelObj
)
}
\arguments{
\item{guide_features}{Weights for guide efficiency features to use on initialization.}

\item{gene_effects}{By-gene essentiality parameters.}

\item{sample_effects}{By-sample parameters.}

\item{use_genes}{Vector of gene names to}

\item{write_log}{Function to write to central timestamped log file.}

\item{user_DataObj}{DataObj with experimental data to analyze.}

\item{user_ModelObj}{ModelObj with preprocessed model parameters.}
}
\description{
Given gene and sample parameters, use MLE to obtain guide parameters.
}
