% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSample.R
\name{optSample}
\alias{optSample}
\title{Function optSample}
\usage{
optSample(
  sample_effects,
  gene_effects,
  guide_efficiency,
  use_genes,
  write_log,
  user_DataObj,
  user_ModelObj
)
}
\arguments{
\item{sample_effects}{Sample effects to use on initiailization.}

\item{gene_effects}{By-gene essentiality parameters.}

\item{guide_efficiency}{By-guide efficiency parameters.}

\item{use_genes}{Vector of gene names to use when estimating sample parameters.}

\item{write_log}{Function to write to central timestamped log file.}

\item{user_DataObj}{DataObject with experimental data to analyze.}

\item{user_ModelObj}{ModelObject with preprocessed model parameters.}
}
\description{
Given guide and gene parameters, use MLE to obtain sample parameters.
Sample parameters assumed to be independent, so 1D optimization used.
All samples present in user_DataObj are used.
}
