% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{read_data}
\alias{read_data}
\title{Ingest BP data to prepare for aCe}
\usage{
read_data(x)
}
\arguments{
\item{x}{csv file with two columns: relative time and relative BP}
}
\value{
data2 -- data table for use in calcLocalMax
}
\description{
This function reads in a dataset of points (relative time x relative blood pressure) extracted from a pdf (generated by EndoPat by Itamar)
}
\examples{
read_data("data.csv")
}
