% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultObjClass.R
\docType{class}
\name{ResultObj}
\alias{ResultObj}
\title{Function ResultObjClass}
\value{
ResultObj R6 class object, with gene_essentiality, guide_efficiency, sample_response,
guide_covariates, gene_essentiality_likelihood, gene_essentiality_CI, gene_essentiality_pvalue,
differential_depletion_pvals, mean_var_model_parameters.
}
\description{
Function ResultObjClass

Function ResultObjClass
}
\details{
An R6 object class to present the results of model fitting as produced by optimizeModelParameters.
Not called by user, but rather within optimizeModelParameters.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gene_essentiality}}{Data.table of gene essentiality
for all, test, and ctrl sample panels.}

\item{\code{guide_efficiency}}{Vector of sgRNA feature names used.}

\item{\code{guide_covariate_score}}{Matrix of sgRNA covariates.}

\item{\code{guide_covariate_weight}}{Vector of sgRNA feature weights.}

\item{\code{sample_response}}{Vector of inferred sample scaling factors.}

\item{\code{gene_essentiality_likelihood}}{Data.table of likelihood values
for all, all_null, test, test_null, ctrl, and ctrl_null tests.}

\item{\code{gene_essentiality_low_CI}}{Lower confidence interval of gene essentiality estimates in \code{gene_essentiality}.}

\item{\code{gene_essentiality_hi_CI}}{Higher confidence interval of gene essentiality estimates in \code{gene_essentiality}.}

\item{\code{gene_essentiality_pvalue}}{Empirical p-values for within- and between-panel essentiality; data.table.}

\item{\code{mean_var_model_parameters}}{Vector of inferred mean-variance relationship parameters; model specific.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResultObj$new()}}
\item \href{#method-clone}{\code{ResultObj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create ResultObj.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultObj$new(optObjList, DataObj, ModelObj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optObjList}}{List from optim fit. With null_liklihood, etc. added on.}

\item{\code{DataObj}}{Produced from raw data.  just need guide2gene_map and num_samples though.}

\item{\code{ModelObj}}{Inferred parameteer object, just need mean_var_model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultObj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
