% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResObj.R
\name{writeResObj}
\alias{writeResObj}
\title{Function writeResObj
write results objects}
\usage{
writeResObj(user_ResObj)
}
\arguments{
\item{user_ResObj}{ResObj or list; output from optimizeModelParameters.}
}
\value{
Returns index of written results version.
}
\description{
Function writeResObj
write results objects
}
