% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveSampleScaling.R
\name{deriveSampleScaling}
\alias{deriveSampleScaling}
\title{Function deriveSampleScaling
create init_scaling and dep_scaling. 
Misspecification will cause our 'neutral' model to
actually be depleted, reducing all essentiality scores.
The expected scaling factor is the ratio of read counts to their
corresponding prior, with the median taken across sgrna in a sample.  
This method assumes the median sgrna is nonessential.
This is also approximated by the ratio of the total sample read counts
to the total expected read counts.
If indicated, the scaling factor will be calculated only using negative
control sgrna (recommended).
If the master library is used as the prior, the expected read count is
the product of the master library sgrna frequency and the total number of
infected cells.
If the master library sgrna frequency is not used, the expected frequency for
 each guide is determined by the 
average percent abundance for that guide across all (mean-normalized) initial
samples.}
\usage{
deriveSampleScaling(
  user_DataObj,
  master_freq_dt,
  use_master_library,
  use_neg_ctrl,
  neg_ctrls,
  write_log
)
}
\arguments{
\item{user_DataObj}{all data; a DataObj.}

\item{master_freq_dt}{normalized masterlibrary frequencies (log)}

\item{use_master_library}{Boolean}

\item{use_neg_ctrl}{Boolean}

\item{neg_ctrls}{Vector of gene name strings.}

\item{write_log}{Function to write messages to log.}
}
\description{
Function deriveSampleScaling
create init_scaling and dep_scaling. 
Misspecification will cause our 'neutral' model to
actually be depleted, reducing all essentiality scores.
The expected scaling factor is the ratio of read counts to their
corresponding prior, with the median taken across sgrna in a sample.  
This method assumes the median sgrna is nonessential.
This is also approximated by the ratio of the total sample read counts
to the total expected read counts.
If indicated, the scaling factor will be calculated only using negative
control sgrna (recommended).
If the master library is used as the prior, the expected read count is
the product of the master library sgrna frequency and the total number of
infected cells.
If the master library sgrna frequency is not used, the expected frequency for
 each guide is determined by the 
average percent abundance for that guide across all (mean-normalized) initial
samples.
}
