% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMasterLib.R
\name{getMasterLib}
\alias{getMasterLib}
\title{Function getMasterLib}
\usage{
getMasterLib(countList)
}
\arguments{
\item{countList}{List of data.tables with counts of sequenced masterlibraries.
One masterlibrary per data.table; all replicates within one data.table.}
}
\value{
Data.table with sgrna x masterlibs of log frequencies.
}
\description{
Process list of master libraries, which may be single or multiple files,
which may have one or more replicates.
Currently just averaging over total-count-normalized replicates.
return data.table of sgRNA x masterlibraries, with column names corresponding
to those referenced in sample_masterlib.
}
