% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optFg.R
\name{optFg}
\alias{optFg}
\title{Function optFg: By-gene essentiality optimization.}
\usage{
optFg(
  startEss,
  sampleSubsets,
  sample_effects,
  guide_efficiency,
  geneList,
  write_log,
  user_DataObj,
  user_ModelObj,
  ncpus = NA
)
}
\arguments{
\item{startEss}{Starting gene essentiality.}

\item{sampleSubsets}{Sample groupings into test, ctrl, and 'all' panels.}

\item{sample_effects}{By-sample parameters.}

\item{guide_efficiency}{By-guide efficiency parameter.}

\item{geneList}{Genes to evaluate sequentially in this function call.}

\item{write_log}{Function to write to central log file.}

\item{user_DataObj}{DataObject containing count data to analyze.}

\item{user_ModelObj}{ModelObject with pre-processed model parameters.}

\item{ncpus}{Number of threads to use in parallel optimization; if not 
specified, will be auto-detected.}
}
\description{
Function optFg: By-gene essentiality optimization.
}
