% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeModelParameters.R
\name{optimizeModelParameters}
\alias{optimizeModelParameters}
\title{Function optimizeModelParameters}
\usage{
optimizeModelParameters(
  user_DataObj,
  user_ModelObj,
  fit_sample_parameter = F,
  subset_genes = NA,
  converge_ll = 100,
  max_iter = 10
)
}
\arguments{
\item{user_DataObj}{DataObj with all experimental data.}

\item{user_ModelObj}{ModelObj with simple inferred parameters for model.}

\item{fit_sample_parameter}{User option to fit by-sample parameters.}

\item{subset_genes}{optional list of start & end points of gene indices to subset.}

\item{converge_ll}{Tolerated variation in likelihood to end optimization; default 100.}

\item{max_iter}{Maximum number of optimization iterations, regardless of convergence. Default 10.}
}
\value{
Data.table with Gene name, optimized essentiality,  guide_efficiency, sample_response,
         guide_covariates, gene_essentiality likelihood, gene_ess CI (hessian),; one DT each for
         "all", "test", and "control" conditions, and an optional 4th for differential depletion
         effect sizes.
}
\description{
Given model object and data object, create a results object.
Mainly runs optim over sample (sub)set(s).
in final version, have as part of R6 object which inherits from ModelObj and DataObj;
for debugging, run as stand-alone with R6 as input.
}
