% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNullData.R
\name{removeNullData}
\alias{removeNullData}
\title{Function removeNullData}
\usage{
removeNullData(user_DataObj)
}
\arguments{
\item{user_DataObj}{Class DataObj, to be stripped. Must be modifiable.}
}
\description{
Function to strip out sgRNA & samples with no count data.
Called by default within DataObjClass.  Do not specify test_tumor_subtype_cols
in DataObjClass
until AFTER erroneous samples are removed, or indicing will be messed up.
}
